clean up visibility of treesync module
