+++ openmls/src/group/tests/test_group.rs
@@ -1,4 +1,12 @@
-use crate::{credentials::*, framing::*, group::*, key_packages::*, test_utils::*, tree::*, *};
+use crate::{
+    credentials::*,
+    framing::*,
+    group::{tests::tree_printing::print_tree, *},
+    key_packages::*,
+    test_utils::*,
+    tree::*,
+    *,
+};
 use openmls_rust_crypto::OpenMlsRustCrypto;
 
 #[apply(ciphersuites_and_backends)]
@@ -384,7 +392,7 @@ fn group_operations(ciphersuite: &'static Ciphersuite, backend: &impl OpenMlsCry
 
     // Make sure that both groups have the same public tree
     if group_alice.treesync().export_nodes() != group_bob.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Alice added Bob");
+        print_tree(&group_alice, "Alice added Bob");
         panic!("Different public trees");
     }
     // Make sure that both groups have the same group context
@@ -471,7 +479,7 @@ fn group_operations(ciphersuite: &'static Ciphersuite, backend: &impl OpenMlsCry
 
     // Make sure that both groups have the same public tree
     if group_alice.treesync().export_nodes() != group_bob.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Alice added Bob");
+        print_tree(&group_alice, "Alice added Bob");
         panic!("Different public trees");
     }
 
@@ -526,7 +534,7 @@ fn group_operations(ciphersuite: &'static Ciphersuite, backend: &impl OpenMlsCry
 
     // Make sure that both groups have the same public tree
     if group_alice.treesync().export_nodes() != group_bob.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Alice added Bob");
+        print_tree(&group_alice, "Alice added Bob");
         panic!("Different public trees");
     }
 
@@ -591,7 +599,7 @@ fn group_operations(ciphersuite: &'static Ciphersuite, backend: &impl OpenMlsCry
 
     // Make sure that both groups have the same public tree
     if group_alice.treesync().export_nodes() != group_bob.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Alice added Bob");
+        print_tree(&group_alice, "Alice added Bob");
         panic!("Different public trees");
     }
 
@@ -674,11 +682,11 @@ fn group_operations(ciphersuite: &'static Ciphersuite, backend: &impl OpenMlsCry
 
     // Make sure that all groups have the same public tree
     if group_alice.treesync().export_nodes() != group_bob.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Bob added Charlie");
+        print_tree(&group_alice, "Bob added Charlie");
         panic!("Different public trees");
     }
     if group_alice.treesync().export_nodes() != group_charlie.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Bob added Charlie");
+        print_tree(&group_alice, "Bob added Charlie");
         panic!("Different public trees");
     }
 
@@ -786,11 +794,11 @@ fn group_operations(ciphersuite: &'static Ciphersuite, backend: &impl OpenMlsCry
 
     // Make sure that all groups have the same public tree
     if group_alice.treesync().export_nodes() != group_bob.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Charlie updated");
+        print_tree(&group_alice, "Charlie updated");
         panic!("Different public trees");
     }
     if group_alice.treesync().export_nodes() != group_charlie.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Charlie updated");
+        print_tree(&group_alice, "Charlie updated");
         panic!("Different public trees");
     }
 
@@ -843,11 +851,11 @@ fn group_operations(ciphersuite: &'static Ciphersuite, backend: &impl OpenMlsCry
 
     // Make sure that all groups have the same public tree
     if group_alice.treesync().export_nodes() == group_bob.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Charlie removed Bob");
+        print_tree(&group_alice, "Charlie removed Bob");
         panic!("Same public trees");
     }
     if group_alice.treesync().export_nodes() != group_charlie.treesync().export_nodes() {
-        print_tree(group_alice.treesync(), "Charlie removed Bob");
+        print_tree(&group_alice, "Charlie removed Bob");
         panic!("Different public trees");
     }
