+++ openmls/src/tree/mod.rs
@@ -2,7 +2,6 @@ use crate::ciphersuite::*;
 
 // Tree modules
 // Public
-pub mod errors;
 pub mod sender_ratchet;
 pub use sender_ratchet::SenderRatchetConfiguration;
 
@@ -12,7 +11,6 @@ pub(crate) mod secret_tree;
 
 pub(crate) mod treemath;
 
-pub(crate) use errors::*;
 use openmls_traits::OpenMlsCryptoProvider;
 
 #[cfg(any(feature = "test-utils", test))]
